refactor(entity-notebook): extract pure helpers to sibling module
Pull 10 pure helpers + 5 shared types out of the 2.2k-line
`EntityNotebookLive.tsx` into `entityNotebookLiveHelpers.ts` so the main
file stays focused on orchestration (state + effects + JSX). Re-exports
at the top of `EntityNotebookLive.tsx` preserve the historical import
path used by `EntityNotebookLive.test.tsx`, so zero test changes are
required.

Moved:
  - chipsToPlainText, chipsEqual, isSyncEditableBlock,
    isTriviallyEmptyNotebookBlock, resolvePresenceSelfUserId
  - shouldRefreshAgentNotebookProjection, getNotebookLoadState
  - extractConvexRequestId, parseNotebookMutationError,
    describeNotebookMutationFailure
  - Types: BlockKind, AuthorKind, AccessMode, LiveBlock,
    ParsedNotebookMutationError

Net: EntityNotebookLive.tsx 2262 → 2034 lines (-228).
Tests: 16/16 passing across both notebook test files.
